+++ StabilityMatrix.Avalonia/Helpers/WindowsPrerequisiteHelper.cs
@@ -23,6 +23,8 @@ public class WindowsPrerequisiteHelper : IPrerequisiteHelper
     private readonly ISettingsManager settingsManager;
 
     private const string VcRedistDownloadUrl = "https://aka.ms/vs/16/release/vc_redist.x64.exe";
+    private const string TkinterDownloadUrl =
+        "https://cdn.lykos.ai/tkinter-cpython-embedded-3.10.11-win-x64.zip";
 
     private string HomeDir => settingsManager.LibraryDir;
 
@@ -43,6 +45,9 @@ public class WindowsPrerequisiteHelper : IPrerequisiteHelper
     private string PortableGitInstallDir => Path.Combine(HomeDir, "PortableGit");
     private string PortableGitDownloadPath => Path.Combine(HomeDir, "PortableGit.7z.exe");
     private string GitExePath => Path.Combine(PortableGitInstallDir, "bin", "git.exe");
+    private string TkinterZipPath => Path.Combine(AssetsDir, "tkinter.zip");
+    private string TkinterExtractPath => PythonDir;
+    private string TkinterExistsPath => Path.Combine(PythonDir, "tkinter");
     public string GitBinPath => Path.Combine(PortableGitInstallDir, "bin");
 
     public bool IsPythonInstalled => File.Exists(PythonDllPath);
@@ -223,6 +228,9 @@ public class WindowsPrerequisiteHelper : IPrerequisiteHelper
             pythonPthContent = pythonPthContent.Replace("#import site", "import site");
             await File.WriteAllTextAsync(pythonPthPath, pythonPthContent);
 
+            // Install TKinter
+            await InstallTkinterIfNecessary(progress);
+
             progress?.Report(new ProgressReport(1f, "Python install complete"));
         }
         finally
@@ -233,6 +241,39 @@ public class WindowsPrerequisiteHelper : IPrerequisiteHelper
                 File.Delete(PythonDownloadPath);
             }
         }
+    }
+
+    [SupportedOSPlatform("windows")]
+    public async Task InstallTkinterIfNecessary(IProgress<ProgressReport>? progress = null)
+    {
+        if (!Directory.Exists(TkinterExistsPath))
+        {
+            Logger.Info("Downloading Tkinter");
+            await downloadService.DownloadToFileAsync(
+                TkinterDownloadUrl,
+                TkinterZipPath,
+                progress: progress
+            );
+            progress?.Report(
+                new ProgressReport(
+                    progress: 1f,
+                    message: "Tkinter download complete",
+                    type: ProgressType.Download
+                )
+            );
+
+            await ArchiveHelper.Extract(TkinterZipPath, TkinterExtractPath, progress);
+
+            File.Delete(TkinterZipPath);
+        }
+
+        progress?.Report(
+            new ProgressReport(
+                progress: 1f,
+                message: "Tkinter install complete",
+                type: ProgressType.Generic
+            )
+        );
     }
 
     public async Task InstallGitIfNecessary(IProgress<ProgressReport>? progress = null)