+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 - Added **Outputs page** for viewing generated images from any package, or the shared output folder
 - Added [Stable Diffusion WebUI/UX](https://github.com/anapnoe/stable-diffusion-webui-ux) package
 - Added [Stable Diffusion WebUI-DirectML](https://github.com/lshqqytiger/stable-diffusion-webui-directml) package
+- Added [kohya_ss](https://github.com/bmaltais/kohya_ss) package
 - Added GPU compatibility badges to the installers
 - Added filtering of "incompatible" packages (ones that do not support your GPU) to all installers 
   - This can be overridden by checking the new "Show All Packages" checkbox
@@ -33,6 +34,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 - Fixed Inference popup Install button not working on One-Click Installer
 - Fixed Inference Prompt Completion window sometimes not showing while typing
 - Fixed "Show Model Images" toggle on Checkpoints page sometimes displaying cut-off model images
+- Fixed missing httpx package during Automatic1111 install
 
 ## v2.5.6
 ### Added